+++ Housing Recommender System/templates/plotly_graph.html
@@ -1,5 +1,209 @@
 <!doctype html>
 <html>
+<head>
+  <title>Recommendations</title>
+  <style>
+
+        body,
+        h1, h2, h3, legend {
+            margin: 0;
+            padding: 0;
+        }
+
+
+        table {
+            border-collapse: collapse;
+            border-spacing: 0;
+        }
+
+        body {
+            font-family: "Roboto", "Arial", "Helvetica Neue", sans-serif;
+            font-weight: 400;
+            font-size: 14px;
+        }
+
+        h1, h2 {
+            font-weight: 400;
+        }
+
+        h1 {
+            font-size: 36px;
+        }
+
+        h2 {
+            font-size: 30px;
+        }
+
+        /* ==========================================================================
+        #FIT INPUT BOX IN CENTER
+        ========================================================================== */
+        .wrapper {
+            padding-top: 70px;
+            margin: 0 auto;
+        }
+
+        .wrapper--w680 {
+            max-width: 680px;
+        }
+
+        /* ==========================================================================
+        #TITLE
+        ========================================================================== */
+        .title {
+          margin-bottom: 10px;
+          font-size: 30px;
+          margin:5px;
+          margin-right:150px;
+          font-family:arial;
+        }
+
+        /* ==========================================================================
+        #CARD FOR ENTERING USER PREFERENCES
+        ========================================================================== */
+        .card {
+              overflow: hidden;
+              -webkit-border-radius: 3px;
+              -moz-border-radius: 3px;
+              border-radius: 20px;
+              background: #FFFFFF;
+              box-shadow: 0 15px 25px rgba(129, 124, 124, 0.2);
+              backdrop-filter: blur(3px);
+              background-color: rgba(155, 155, 155, 0.3);
+        }
+
+
+
+        .card-1 .card-body {
+             padding: 0 70px;
+             padding-top:  10px;
+             padding-bottom: 15px;
+             background: #1a1a1a;
+             box-shadow: 0 15px 25px rgba(129, 124, 124, 0.2);
+             backdrop-filter: blur(3px);
+             background-color: rgba(255, 255, 255, 0.2);
+        }
+        /* styling navlist */
+		#navlist {
+			width: 100%;
+		}
+
+		/* styling navlist anchor element */
+		    #navlist a {
+			float:left;
+			display: block;
+			color: white;
+			text-align: center;
+			padding: 12px;
+			text-decoration: none;
+			font-size: 15px;
+		}
+		.navlist-right{
+			float:right;
+		}
+
+		/* hover effect of navlist anchor element */
+		#navlist a:hover {
+			background-color: #ddd;
+			color: black;
+		}
+
+        /* ==========================================================================
+        #BACKGROUNG IMAGE
+        ========================================================================== */
+
+        :root {
+            --gap-line-spacing: 5px;
+            --gap-line-spacing-authors: 2px;
+            --darkest-color-titles: #545454;
+            --main-text-color: #003057;
+            --light-color-section-banner-bkg: #B3A369;
+            --top-banner-color: #B3A369;
+            --gap-line-spacing-company: 2px;
+            --page-background-color: #ffffff;
+            --table-border-color: #B3A369;
+            --section-banner-text-color: #003057;
+            --font-my-theme: Helvetica;
+            --gap-right-before-links: 2em;
+        }
+
+        body, html {
+            height: 100%;
+            margin: 0;
+        }
+
+        .bg {
+            background-image: url('/static/House.png');
+            background-position: center-center;
+            background-size: cover;
+            background-repeat: no-repeat;
+            height: 100%;
+        }
+
+        .title {
+            color: var(--main-text-color);;
+            font-family: Arial, Helvetica, sans-serif;
+            font-size: 50px;
+        }
+        .container .box {
+            display: flex;
+            flex-direction: row;
+        }
+
+            .container .box .box-cell.box1 {
+                width: 80%;
+                background: white;
+                color: black;
+                text-align: justify;
+                text-shadow: 0.1px 0.1px 0.1px #000;
+                backdrop-filter: blur(0px);
+                background-color: rgba(255, 255, 255, 0);
+            }
+
+            .container .box .box-cell.box2 {
+                width: 30%;
+                background: white;
+                text-align: justify;
+                text-shadow: 0.1px 0.1px 0.1px #000;
+                backdrop-filter: blur(0px);
+                background-color: rgba(255, 255, 255, 0);
+            }
+
+        .title_label {
+            color: var(--main-text-color);
+        }
+
+        .imp_td {
+            text-align: center;
+            padding-left: 10pt;
+            width:35%;
+        }
+
+        .imp_td_text {
+            text-align: left;
+            padding-left: 1em;
+            width: 10%;
+        }
+        .image {
+             background-size: 500px;
+             background-color: transparent;
+             width: 500px;
+             height: 90px;
+             position: absolute;
+             top:2%;
+             right: 33%;
+       }
+       iframe {
+       width: 700px;
+       height: 500px;
+       margin: 0 auto;
+       display: block;
+       border-style:none;
+       background:#777;
+       background-color: transparent;
+    }
+
+    </style>
+</head>
 <body style="font-family:arial, sans-serif">
     <h1>{{header}}</h1>
     <h6><a href="/">Return to index page</a></h6>
